[Refactoring] Move modles from EpoxyModels.kt files to .ui.models package (also fix #11)
